changed int typecast to float incase user inputs a decimal

+++ week-2/assignment-2/mini_project.py
@@ -1,4 +1,4 @@
-fahrenheit = int(input("Enter a temperature in Fahrenheit: "))
+fahrenheit = float(input("Enter a temperature in Fahrenheit: "))
 
 celsius = (fahrenheit - 32) * 5/9
